Make python bindings use threads.
The index operation can now take a number of threads to use. The default,
if none is given, is to use the number of CPUs. To specify a number of
threads pass it as the only argument to make_index().

Move mi_data into common.h and out of snugglefish.h.

Various whitespace cleanup in pysnugglefish.