Make search take an optional thread argument.
When searching using pysnugglefish you can now provide an optional
thread argument. If not provided the default is to use the number of CPUs.

Also, if the threads is <= 0 raise an exception instead of defaulting. This
is now done for both searching and indexing.

Make snuggle.py default to number of CPUs is no thread option is provided.
I could make this not pass an argument but it's less clean to do that.

While here, add exception handling in snuggle.py.